+++ resources/views/members/show.blade.php
@@ -57,6 +57,9 @@
     @if($mainUser && $canViewDevices)
     <a class="m-btn" href="{{ route('devices.by_user', $mainUser->id) }}">Zařízení</a>
     @endif
+    @if($canViewConnectionRequests)
+    <a class="m-btn" href="{{ route('connection_requests.by_member', $member->id) }}">Žádosti o připojení</a>
+    @endif
     @if($canViewInvoices)
     <a class="m-btn" href="{{ route('invoices.by_member', $member->id) }}">Faktury</a>
     @endif